LayoutTests/fast/scrolling/mac/clip-path-hit-test-expected.txt


Test scroll over content
PASS overflowScrollEventCount > 0 is true
PASS windowScrollEventCount is 0

Test scroll over content
PASS overflowScrollEventCount is 0
PASS windowScrollEventCount > 0 is true

Test scroll over content
PASS overflowScrollEventCount is 0
PASS windowScrollEventCount > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
